backend/requirements.txt
fastapi
uvicorn
gitpython
pydantic
python-multipart
requests
